Added Note Drawing

+++ Assets/Scripts/BmsParser/StopHeader.cs
@@ -1,11 +1,11 @@
 ﻿class StopHeader
 {
-    public int id; //id of file
+    public ulong id; //id of file
     public ulong duration;
 
     public StopHeader(int id, ulong duration)
     {
-        this.id = id;
+        this.id = (ulong)id;
         this.duration = duration;
     }
 }
